Wolfram Sang <wolfram@the-dreams.de>
I think a small HISTORY file could be useful (like "when again was
this feature introduced?"). As such information is already on the
webpage, I attached a proposal (simply webpage dumped with elinks and
removed the unneccessary stuff).

Me:

Named the file ChangeLog as this seems to be the convention with
many GNU projects (although we aren't one :)