Antonio Borneo <borneo.antonio@gmail.com>
I have realized that the official Netlock/Contivity client converts
the group ID string to lowercase before sending it in the first ISAKMP
packet. (... to be more accurate, it sends an hash on a set of data
including the group ID).
I have found no usefule info on Internet. Instead, plenty of sites
report that group ID is case sensitive!

After changing to lowercase the group ID in my vpnc config file I was
able to go ahead as with the other account.

Is there any of you that uses vpnc-nortel with uppercase characters in group ID?

Could you please verify my statement on your systems?
Try to change to uppercase one character in group ID in either
Netlock/Contivity and vpnc clients, then login in your Nortel server.
Netlock/Contivity should be able to login (since it will change the
group ID to lowercase), while vpnc should fail at first packet.
If this behaviour is confirmed, vpnc-nortel should only use lowercase
group ID, or convert it to lowercase, or at least rise a warning to
the user.

Testing more, l can summarize:
Group ID: Nortel client convert to lowercase before use it
Group PWD: case sensitive
User name: client sends it with proper case, but (at least my) server
is case insensitive
User pwd: case sensitive

 ###

There is a mistake in my previous mail.
I wrote :
> ... it sends an hash on a set of data including the group ID).
It is not on "a set of data"; it is directly the SHA1 hash of group
ID, sent as data payload in the ISAKMP_PAYLOAD_ID.
With Wireshark you can extract such data from the first packet sent,
and compare it with this offline computation
# echo -n "my group ID" | sha1sum
and verify that the client converts to lowercase before the hash.

I confirm that both Windows and Linux version of Contivity client
convert Group ID to lowercase before computing SHA1 hash