Jukka Salmi <j+vpnc@2009.salmi.ch>
on systems without an error(3) function (i.e., where HAVE_ERROR is not
defined), making the test-crypto target currently fails because crypto.c
uncondinionally includes <error.h> (and that header does not exist on
such systems), and because sysdep.o (the file containing the missing
error() function) is not specified as a source for the target in
question.

The attached patch fixes the problem (compile-tested on NetBSD/i386).